excel-chart-generator/requirements.txt
streamlit==1.40.0
pandas==2.2.3
numpy==1.24.4
openpyxl==3.1.5
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2
google-generativeai==0.8.3
python-dotenv==1.0.1